web build fixes
