+++ babel.config.js
@@ -1,5 +1,6 @@
 module.exports = function(api) {
   api.cache(true);
+
   return {
     presets: [['module:metro-react-native-babel-preset']],
     env: {
