+++ src/components/dashboard/FeedItems/ListEventItem.js
@@ -5,7 +5,7 @@ import { isMobile } from '../../../lib/utils/platform'
 import normalize from '../../../lib/utils/normalizeText'
 import { getFormattedDateTime } from '../../../lib/utils/FormatDate'
 import { withStyles } from '../../../lib/styles'
-import { getScreenWidth } from '../../../lib/utils/Orientation'
+import { getScreenWidth } from '../../../lib/utils/orientation'
 import Avatar from '../../common/view/Avatar'
 import BigGoodDollar from '../../common/view/BigGoodDollar'
 import Text from '../../common/view/Text'
