+++ src/components/dashboard/FeedModalList.js
@@ -5,7 +5,8 @@ import { Portal } from 'react-native-paper'
 import { once } from 'lodash'
 import { isMobileOnly } from '../../lib/utils/platform'
 import { withStyles } from '../../lib/styles'
-import { getMaxDeviceWidth, getScreenHeight, getScreenWidth } from '../../lib/utils/Orientation'
+import { getScreenHeight, getScreenWidth } from '../../lib/utils/orientation'
+import { getMaxDeviceWidth } from '../../lib/utils/sizes'
 import { CARD_SLIDE, fireEvent } from '../../lib/analytics/analytics'
 import FeedModalItem from './FeedItems/FeedModalItem'
