+++ src/components/signin/MnemonicInput.js
@@ -4,7 +4,7 @@ import { StyleSheet, TextInput, View } from 'react-native'
 import { values } from 'lodash'
 import logger from '../../lib/logger/pino-logger'
 import { withStyles } from '../../lib/styles'
-import { getScreenHeight } from '../../lib/utils/Orientation'
+import { getScreenHeight } from '../../lib/utils/orientation'
 import Text from '../common/view/Text'
 
 const height = getScreenHeight()
