+++ src/components/signup/PhoneForm.js
@@ -4,7 +4,7 @@ import { debounce } from 'lodash'
 import { isMobile } from '../../lib/utils/platform'
 import { getDesignRelativeHeight } from '../../lib/utils/sizes'
 import { userModelValidations } from '../../lib/gundb/UserModel'
-import { getScreenHeight } from '../../lib/utils/Orientation'
+import { getScreenHeight } from '../../lib/utils/orientation'
 import logger from '../../lib/logger/pino-logger'
 import SimpleStore from '../../lib/undux/SimpleStore'
 import { withStyles } from '../../lib/styles'
