+++ src/lib/utils/orientation.js
@@ -1,6 +1,5 @@
 import { Dimensions } from 'react-native'
-import { isBrowser } from '../utils/platform'
-import { theme } from '../../components/theme/styles'
+import { isBrowser } from './platform'
 
 let originalScreenHeight = 0
 
@@ -16,22 +15,8 @@ if (!originalScreenWidth) {
 
 export const getOriginalScreenHeight = () => originalScreenHeight
 export const getOriginalScreenWidth = () => originalScreenWidth
-
 export const getScreenHeight = () => Dimensions.get('window').height
-export const getMaxDeviceHeight = () => {
-  const height = getScreenHeight()
-  const { maxHeightForTabletAndDesktop } = theme.sizes
-
-  return Math.min(height, maxHeightForTabletAndDesktop)
-}
-
 export const getScreenWidth = () => Dimensions.get('window').width
-export const getMaxDeviceWidth = () => {
-  const width = getScreenWidth()
-  const { maxWidthForTabletAndDesktop } = theme.sizes
-
-  return Math.min(width, maxWidthForTabletAndDesktop)
-}
 
 export const isPortrait = () => {
   return isBrowser ? true : getScreenHeight() >= getScreenWidth()
