+++ src/lib/utils/sizes.js
@@ -1,4 +1,21 @@
-import { getMaxDeviceHeight, getMaxDeviceWidth, isPortrait } from './Orientation'
+import { theme } from '../../components/theme/styles'
+import { getScreenHeight, getScreenWidth, isPortrait } from './orientation'
+
+// have moved the following two functions here to break the circle dependency
+// components/theme/styles -> lib/utls/normalizeText -> lib/utils/orientation -> components/theme/styles
+export const getMaxDeviceWidth = () => {
+  const width = getScreenWidth()
+  const { maxWidthForTabletAndDesktop } = theme.sizes
+
+  return Math.min(width, maxWidthForTabletAndDesktop)
+}
+
+export const getMaxDeviceHeight = () => {
+  const height = getScreenHeight()
+  const { maxHeightForTabletAndDesktop } = theme.sizes
+
+  return Math.min(height, maxHeightForTabletAndDesktop)
+}
 
 /**
  * Receives a size matching the designs baseSize and converts to dp on the current device
@@ -11,6 +28,7 @@ const getDesignRelativeSize = (size, isMax = true, baseSize, currentSize) => {
   const sizeInVW = size / baseSize
   const relativeSize = currentSize * sizeInVW
   const calculatedSize = isMax ? Math.min(size, relativeSize) : relativeSize
+
   return calculatedSize
 }
 
@@ -23,6 +41,7 @@ const DESIGN_WIDTH = 360
  */
 export const getDesignRelativeWidth = (width, isMax = true) => {
   const screenWidth = isPortrait() ? getMaxDeviceWidth() : getMaxDeviceHeight()
+
   return getDesignRelativeSize(width, isMax, DESIGN_WIDTH, screenWidth)
 }
 
@@ -35,5 +54,6 @@ const DESIGN_HEIGHT = 640 - 24
  */
 export const getDesignRelativeHeight = (height, isMax = true) => {
   const screenHeight = isPortrait() ? getMaxDeviceHeight() : getMaxDeviceWidth()
+
   return getDesignRelativeSize(height, isMax, DESIGN_HEIGHT, screenHeight)
 }
